swapped manual dynamic chunked loader with readt-loadable
