+++ client/coral-embed-stream/src/tabs/configure/components/loadable/QuestionBoxBuilder.css
